Fix cluster.rc tests with a hack for now (#4508)
On the latest version of ubuntu all tests that have cluster.rc are
failing. Mostly because of this:
https://datatracker.ietf.org/doc/html/rfc6890

The following patch fixes it for now. But we may need to change the
framework to fix it properly I think. Ideally localhost or loopback ip
range shouldn't be allowed as hostname. This part we may have to address
separately.

> Fixes: #4457
> Change-Id: Iee3e6710c301cde4d5bb250ff45fd84031a878be

Change-Id: Iafba1faeee52cc453f8c59cbc34fc8f1c80412fb
backport-of: 0707922ab630d9b7b5d33ee9568b2a4a5fa36790
Signed-off-by: Pranith Kumar Karampuri <pranith.karampuri@phonepe.com>