tests: address XFS 300M minimum size (#4361) (#4487)
* tests: shared-statfs.t XFS 300M minimum (#4361)

	A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: bug-1091935-brick-order-check-from-cli-to-glusterd.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: df-results-post-replace-brick-operations.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: bug-1586020-mark-dirty-for-entry-txn-on-quorum-failure.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: bug-948729-force.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: bug-948729.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: bug-948729-mode-script.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: distribute/file-create.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: arbiter-statfs.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: zero-fill-enospace.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: issue-1254-prioritize-enospc.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests: fuse/bug-858488-min-free-disk.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* tests:  tests/bugs/distribute/bug-1099890.t XFS 300M minimum (#4361)

        A commit in xfsprogs raised the minimun size of a xfs filesystem
to 300M, causing the tests to fail because they were trying to build
even smaller filesystems. This commit adjust the size parameters to fit.

Signed-off-by: Thales Antunes de Oliveira Barretto <thales.barretto.git@gmail.com>

* Update tests/bugs/fuse/bug-858488-min-free-disk.t

Co-authored-by: Yaniv Kaul <yaniv.kaul@scylladb.com>]

Change-Id: I76e449fda1892dec15021ddfa354e32daadebb13
backport-of: 97071145669b22ea1030dc65b1baa68b91433079
Signed-off-by: Pranith Kumar Karampuri <pranith.karampuri@phonepe.com>