linkfile: use fops->create in dht_linkfile_create
A newly created linkfile will be deleted for the race, after
this, the gfid between linkfile and datafile may get
mismatched. And shards deletion process is also affected
by such a mismatch, most shards in BRICK_PATH/.shard
cannot be removed correctly and space cannot be released.

To fix this race, replace mknod with create during linkfile
creation, in this case, we can keep this linkfile opening
during the whole linkfile and datafile creations.

A linkfile will not be regarded as stale and deleted then,
because its fd is opened and posix_unlink can handle it
properly.

In this patch, the linkfile and datafile creation process is as below:

    dht                       hashed_subvol                cached_subvol
--> |                              |                              |
    |                              |                              |
    | a.create tmp fd as local->fd |                              |
    | b.send fops->create          |                              |
    |---- dht_linkfile_create ---->|                              |
    |                              |                              |
    | c.succeed to create linkfile,|                              |
    |<-- dht_linkfile_create_cbk --|                              |
    |                                                             |
    |--- corresponding fop for datafile, mknod, create, etc.  --->|
    |                                                             |
    |<----------            datafile fop cbk            ----------|
    |                                                             |
    |                              |                              |
    | d.stack unwind/destroy       |                              |
    | e.clear dht_local_t in       |                              |
<-- |   dht_local_wipe             |                              |
    |                              |                              |

A tmp local->fd will be created for non-fd-based fops to ensure
safe access to fd or its member in subsequent procedures. Once this
linkfile was successfully created, its fd_count will be increased in
server cbk and passed to client if necessary. Then, between phase c
and phase e, linkfile will not be deleted by mistake since later
processes will know this linkfile is being used, instead of stale.
After phase e, this linkfile links to a datafile, so it is valid.

This change should also be applied during dht_lookup_everywhere_cbk.

Fixes: #4548
Signed-off-by: chenjinhao <chen.jinhao@zte.com.cn>